refactor: Moved `mergeXmlWorkflow` to presentation.ts Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
